requirements.txt
asgiref==3.8.1
click==8.2.1
colorama==0.4.6
cssbeautifier==1.15.4
Django==5.0.2
django-browser-reload==1.18.0
django-filter==23.5
django-tailwind==3.8.0
djlint==1.36.4
EditorConfig==0.17.0
Faker==37.3.0
jsbeautifier==1.15.4
json5==0.12.0
pathspec==0.12.1
pillow==10.2.0
PyYAML==6.0.2
regex==2024.11.6
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
